requirements.txt
pandas==1.3.5
psycopg2-binary==2.9.5
requests==2.27.1
python-dotenv==0.21.1
apache-airflow==2.3.3